llmps_model/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
matplotlib>=3.3.0
pillow>=8.0.0
tqdm>=4.50.0
tensorboard>=2.4.0
scikit-learn>=0.24.0
transformers>=4.5.0
pywt>=1.1.0
PyWavelets>=1.1.0 